fix: harden analysis pipeline against crashes, leaks, and degenerate splits
- vmem_utils: split_boundary now snaps to the nearest INTERIOR sequence
  edge; single-sequence runs fall back to the ratio cut instead of
  degenerating to a 1-frame eval set (test pipeline was 1191/1192)
- severity: read fitted representation from split.json, skip
  flow_pca.joblib, precompute detector clean scores once, and guard
  scoring so one bad run cannot abort the whole analysis
- evaluate_detectors: skip runs whose feature width does not match the
  fitted detectors instead of crashing mid-benchmark
- extract_ann_baselines: atomic temp-file+rename saves so interrupted
  writes cannot poison the exists()-based resume check
- evaluate_ann_baselines: ignore leftover _tmp_ files when globbing runs
- representation_ablation: keep phi-only runs (warn instead of dropping
  them from all downstream stages); fit Mahalanobis once per
  representation instead of refitting per run (~270 fewer O(d^3) fits)
- vmem_models: RealNVP log-det accumulator inherits input dtype
- add run_analysis_tests.py: 31-test regression suite for analysis/

+++ analysis/vmem_models.py
@@ -45,7 +45,7 @@ class RealNVP(nn.Module):
         self.layers = nn.ModuleList([CouplingLayer(dim, hidden_dim) for _ in range(n_layers)])
         
     def forward(self, x):
-        log_det_tot = torch.zeros(x.shape[0], device=x.device)
+        log_det_tot = torch.zeros(x.shape[0], device=x.device, dtype=x.dtype)
         z = x
         for i, layer in enumerate(self.layers):
             z, log_det = layer(z, flip=(i % 2 == 1))
